Move global functions into objects
